Pycharm! That was the worst automatic refector I've ever seen!
