+++ datacube_ows/styles/component.py
@@ -1,7 +1,6 @@
 from xarray import Dataset, DataArray
 import numpy as np
 
-import tests.utils
 from datacube_ows.ogc_utils import ConfigException, FunctionWrapper
 from datacube_ows.styles.base import StyleDefBase
 
@@ -107,7 +106,7 @@ class ComponentStyleDef(StyleDefBase):
                         imgband_data = imgband_component
                 if imgband_data is None:
                     imgband_data = np.zeros(list(data.dims.values()), 'uint8')
-                    imgband_data = DataArray(imgband_data, tests.utils.coords, data.dims.keys())
+                    imgband_data = DataArray(imgband_data, data.coords, data.dims.keys())
                 if imgband != "alpha":
                     imgband_data = self.compress_band(imgband, imgband_data)
                 imgdata[imgband] = imgband_data.astype("uint8")
