+++ datacube_ows/styles/ramp.py
@@ -10,8 +10,6 @@ from matplotlib import pyplot as plt
 from matplotlib.colors import to_hex, LinearSegmentedColormap
 from xarray import Dataset
 
-import tests.utils
-
 matplotlib.use('Agg')
 
 from datacube_ows.ogc_utils import ConfigException, FunctionWrapper
@@ -503,7 +501,7 @@ class ColorRamp:
         imgdata = {}
         for band in self.components:
             imgdata[band] = (data.dims, self.get_8bit_value(data, band))
-        imgdataset = Dataset(imgdata, coords=tests.utils.coords)
+        imgdataset = Dataset(imgdata, coords=data.coords)
         return imgdataset
 
     def color_alpha_at(self, val):
